test/inspector/regress/regress-crbug-1270780-expected.txt

Regression test for crbug.com/1270780.

Running test: testDontBreakOnCaughtException
SUCCESS
